+++ data/quest/fabrication_station_definition.gd
@@ -1,0 +1,5 @@
+class_name FabricationStationDefinition
+extends Resource
+
+@export var station_id := "station_fabrication"
+@export_range(0.1, 30.0, 0.1) var fabrication_duration_seconds := 2.5